Commit: Fixing bugs in lattice file
For the demo, I found some bugs in the lattice file. I have commented
some lines in the adaptor, so that the lattice file will print awd as
aw0 instead of the internal calculated value. Also, in order to fix the
f1st bug, I modified the read input function in order to write the
xlamds from the GENESIS input file and not calculating it from the fel
parameters function.