Lattice file : Bug when a GENESIS file is present
A bug in the aw0 parameter was fixed in the script, when the GENESIS
file is chosen to be set to be 1. The aw0 which is read from the GENESIS
file should be divided by \sqrt{0.5} (GENESIS multiplies the K undulator
factor by \sqrt{0.5} when the lattice file is generated). The
distinction between helical and planar undulator comes when the user
does not have a GENESIS file. The script calculates the K undulator
factor and does the rescaling by /sqrt{2} depending on whether it is
helical or planar