Fixing the bug of the time window
Dave reported a bug concerning the nslice and ntail. When the class is
called, OCELOT automatically calculates nslice from the curlen (an
attribute of the beam object) and sets the ntail to 0. Now, if there is
an existent beam file, the class will calculate nslice as OCELOT does,
and sets ntail to 0. If there is not an existent beam file, the class
will read the nslice from the input file and sets ntail to -nslice/2.

The nslice and ntail parameters can be overwritten by using the
GEN_rewrite_par method of the class. The only thing that the user needs
to do is to set the i_rewrite flag in the dictionary as 1 and write the
parameters to be rewritten together with their new values (as a
dictionary within the dictionary). See the script in the repository for
a way to implement it