Fixing of bugs (reading from input file ) + normal plot of power curve (_rad.png)
This commit solves several bugs:
- gui/genesis_plot.py: The plot that is created for each noise
realisation within the root directory(__rad__.png) now shows the power
in a nonlogarithmic scale
- S2E_STFC/FEL_simulation_block.py:
* If one of the specs of the input file does not start with a space (not
aligned), the reading method adds an space in order to align it with the
other parameters, so that the parameter can be read
* Once you have set up the input object via the generate_input, the
class now overwrites those values by the ones from the input file (this
is done in order to recover some of the simulation parameters).
However, in the case of the rxbeam, rybeam, alphax, alphay, those
parameters are calculated from the beam object created before.
* Also, in order to guarantee that the simulation is time dependent, the
attribute type of the input object is set to be tdp if itdp =0 (this is
done since the input object was overwritten)