Commit: Fixing lattice file generation (genesis.py and class), beam file plotting (genesis.py)
- The following bugs have been fixed :
* Generation of the lattice file (now if f1st =1, the lattice file will
be made with the 1st aw shifted corresponding to the 1st half of fodo
cell).
* Bug of plotting the existent beam file: If there is no I or eloss
column, the plot_beam function does not raise an error (a workaround was
done)
* Adding new extraflags to the script that calls the class : It shows
all the different entries of the dictionary. All you have to do is set
the right paths or flags depending on the scenario