Commit: Bug fix (centering of ASTRA distribution, quad scan fixed and plotting of slice properties)
The following fixes have been implemented:

- The quad scan has been fixed (now , it should work as intended. There
is still a relative difference less than 10% between the values given by
betamatch when the function is called and doing it as stand alone. Could
be due to the way that the lattice structure has been defined). It can
find the solution as expected for different values of quads which the
solution is found when betamatch is run.

- Now the script does not fail when the input file has a line of
beamfile or distfile.

- When the betamatch funcion is called, a statement saying that the
parameters have changed or not is displayed (before, it used to say :
'Betamatch did not work ' or 'Betamatch worked', which was confusing
since it could be that it did work but the twiss parameters were not
changed.)

- Most important fix: When the distribution is read from an ASTRA file,
the distribution is centered when converted to an edist object.

- Postprocessing: When a distfile object exists (read from ASTRA or read
from existent edist file), in order to get the properties per slice, the
object is converted into a beam object via the function edist2beam in
OCELOT. Afterwards, the object is plotted via the plot_beam function.
The file with the slice information is saved as slice_edist.png within
the root directory.