Taper scan
The daughter class FEL_taper has been implemented. It allows the user to
scan over linear tapers, starting from a module within the undulator. It
worked also with the i_rewrite flag, in order to set up a taper scan in
the case of a seeded scenario

The input dictionary has two additional flags: i_taper(which enables the
taper optimisation), and par_taper (a dictionary which contains the
n_starting entry (tuple of starting modules of the taper), num (number
of tapers), a_tmin (linear coefficient of the initial taper), a_tmax
(linear coefficient of the last taper of the scan).

The user will get a tree structure: The untapered results are saved
within the untapered subdirectory. The tapered results will be saved
within the taper_scan subdirectory. There is a subfolder created for
each possible combination (for example for a taper which starts at the
7th module and with linear coefficient 0.001):
taper_scan/n_start_7/tap_lin_-0.001/scan_0/ip_seed_-1/

The class generates the lattice file with the taper function applied to
it during running time.

At the end of the taper scan, two plots are generated within the
taper_scan folder:
- pulse_energy_tapering.png (Pulse energy curves for each taper
scenario. The untapered case is represented by a green dotted thick
line).
- relative_difference_tapering: Relative difference (difference of max
pulse energy between taper and untapered over max pulse energy
untapered) vs taper coefficient. Each curve represents a starting
module. This allows the user to identify the optimum taper.

The taper optimisation is done in time-dependent mode.