Ming-Xie class, moga preparations and after-burner preparations
- Ming Xie Class: Calculates the pulse energy, rho parameter, saturation
power, gain length using the Ming Xie formalism. The class has several
attributes for the calculation of the FEL parameters using the Ming Xie
formalism (output - where the results are stored, iplot - flag
corresponding to the generation of a plot with the results in the
output/mXie folder, npartmax - total number of particles of the
distribution, fl, drl ,dl, K and nwig - undulator and FODO design
parameters which can be read from a GENESIS input file , wavelength -
radiation wavelength, lambdu - undulator period, binwidth - bin size).
The class has several methods which calculates the rho parameter, gain
length, saturation power, saturation length, noise power per slice. The
one that runs the calculations is ming_xie_calculation

- adaptors/genesis.py : Modifications to the adaptor in order to run the
afterburner (generation of beam file only including the energy and
energy spread). Modification of the lout attribute of the input object
with the new format of the input file.
-cpbd/moga.py: Modifications for the MOGA run (generating the results
for the generations and not overwriting the binary file.
-gui/genesis_plot.py: small modification in order to have the electron
beam sized weighted average by the current
- S2E_STFC/FEL_simulation_block.py: Modification to the Undulator
automatic generation of lattice file so that it generates the lattice
file only if f1st =1. Otherwise, it will read it from the input file.
Modifications done in order to set the afterburner