Bug fixing (lattice file) and afterburner (current in beam file)
A bug was identified in the generation of the lattice file, which makes
the undulator modules shorter, when f1st =1 in the input file. So, the
fixed has been implemented to the generate_lattice function in such a
way that the 1st line of the AD (virtual drift) is shifted by one
undulator period in order to have the correct matching.

Also, the current column has been added to the beam object in the case
of the afterburner(edist2beam and  beam_file_str methods of the
genesis.py adaptors script)