Update: Running the FEL simulation block if a lattice file is present and bug in genesis.py (edist2beam)
The FEL simulation block was modified in order to allow the user to run
a GENESIS simulation with an existent lattice file.

The lattice file should be in the directory where the GENESIS input file
is located.

Also a bug in the edist2beam routine is fixed so that there are no
issues with the slice properties plot